requirements.txt
pyTelegramBotAPI==4.23.0
psycopg2-binary==2.9.9 